tdd: support with VS202x
The following modification has been add:
- SYS_SLIST_FOR_EACH_CONTAINER_SAFE_WITH_TYPE
- TYPE_SECTION_FOREACH (custom version)
- _arch_isr_direct_pm (custom version)
- LOG2_WITH_TYPE, __z_log2_with_type
- __builtin_clzll

+++ kernel/init.c
@@ -503,7 +503,7 @@ void __weak z_early_rand_get(uint8_t* buf, size_t length) {
 
     if ((entropy != NULL) && device_is_ready(entropy)) {
         /* Try to see if driver provides an ISR-specific API */
-        rc = entropy_get_entropy_isr(entropy, buf, length, ENTROPY_BUSYWAIT);
+        rc = entropy_get_entropy_isr(entropy, buf, (uint16_t)length, ENTROPY_BUSYWAIT);
         if (rc > 0) {
             length -= rc;
             buf += rc;
